Improve the warning for missing backend env vars
The previous message was slightly confusing and could cause people to think there was a security issue when the backend keys are missing. This could be the case but its also true that some apps are ok to have backend authentication disabled and in these cases people should update the `withBackendAuthentication` so the middleware never runs.