FTDCS-37 Fix the restart check tests
This addresses the test failures introduced in #675, which were caused
because we were expecting only a single health check to be exported by
the `health-checks` module.

After a discussion with Serena and Ivo, we decided to refactor the
`health-checks` module to export and array of checks, and update
`InstrumentListen#addMetrics` to accept either an array or a single
item.