Switch from graphiteSpike to graphiteThreshold for app restarts
I think this threshold check will be more reliable than looking for spikes.

The way this check works is a little confusing at first but very good
for us:

- The wildcard matcher (for example *_EU) will match all dynos for
  a given app
- Graphite will return multiple series for this query (dyno_1_EU,
  dyno_2_EU, etc)
- The graphiteThreshold check will iterate over all these series
  and check if any of them are above a given value
- I have used the summarize function to group these series into
  1 hour buckets, and transformNull to set null data (no restarts) to 0
- The samplePeriod means that we go back 6 hours (6 metrics per dyno)

I reckon if a dyno restarted more than 2 times in a one hour period that
would be a good indication of a problem.