Add express.urlencoded to exports
Express version 4.16+ has its own body-parser implementation included in the default Express package so there is no need to add another dependency.
The 2 most used functions of body-parser along the projects are json and urlencoded.
Adding urlencoded to the exports of n-express allows us to remove the body-parser dependency from all (or at least most) of the projects.