Add a default Heroku log drain health check
This uses the new Heroku log drain check which allows us to see alerts
for an app if it's expecting to use log drains but they're misconfigured
in some way. This default check is only enabled if an app has the
`MIGRATE_TO_HEROKU_LOG_DRAINS` environment variable set so it won't
start failing for apps which haven't conciously opted into log drains
(See https://github.com/Financial-Times/n-logger/pull/152 for more info).

There's a note in the source explaining that this condition will be
removed in future. I think this is a minor change as the health check
will only start failing when an app opts in.

Relates-To: FTDCS-218