Require Heroku log drains in all n-express apps
Now that we have a timeline for completing the log drain migration
(30th June 2023), we want to start failing health checks for apps which
don't have them configured yet. The condition based on the migration
environment variable was always intended to be removed (FTDCS-233) and
now feels like a good time.

Resolves FTDCS-316.