Honour the error status code
We have useful information across all our apps which we're just
ignoring. We use HTTP errors extensively which set either a `status` or
`statusCode` property and there's no way for these errors to actually
impact the status code which is sent to the end-user. This means that
our Fastly logging isn't particularly useful, e.g. we can't say:

> Hmm, Fastly is logging a lot of `502` errors, I wonder which upstream
> service is causing us issues.

I don't think that this was ever given lots of thought as it was added 5
years ago without much of a description: 3f0f2af

Does anyone object to us honouring an error's status code if it has one?