Correctly handle ERR_STREAM_WRITE_AFTER_END errors
We got an app crash alert in the ft-next-support channel today where
next-syndication-api was sending a response after one has already been
sent. I'm certain we need to do some work in that app too, but the error
doesn't have enough detail for me to work out where the response is
being written.

The error highlighted that we don't handle this case particularly well -
the Express documentation suggests that you should always fall back to
the default error handler if headers have already been sent to avoid
this. I tracked this down to the finalhandler module which Express uses
and it correctly terminates the request instead of causing a crash.

I decided to use Reliability Kit's serializers here so that we have the
same amount of detail as in our other errors. The main useful thing for
debugging next-syndication-api will be having a log of the request that
was made so that I can work out which route is erroring.