fix: CI build and test jobs should run on node v14 + v16
As there is not a way in Tool Kit to generate a matrix, we have had to
remove the "CONFIG GENERATED BY DOTCOM-TOOL-KIT" comment in the Circle
CI config. This means we can manually amend the config to support this
but it will mean more effort to add any changes from Tool Kit later. As
we will be migrating all systems to Node 18 in the next year, we do not
see a need to support generating a matrix for Tool Kit generated Circle
CI configs.