feat: Add new Heroku log drain healthcheck
This check uses a new endpoint on next-health which will proxy
the Heroku API and therefore doesn't impact our Heroku API quota
too badly.

This healthcheck will roll out to all applications automatically
as they upgrade to this version of n-express.

We can deploy next-health to adjust the success/failure criteria of
this check instantly.

This commit is a sort-of revert of 19e9ecefc224115cd7592d0.