CircleCi config fix: Remove dependency on node 14 test job for publish
The circle workflow for n-express runs two parallel jobs - a build and test job for Node 16 and a build a test job for Node 14.
Each job attaches and persists a workspace. Previously both jobs were required for the publish job to commence, but this causes CircleCi to error because of the duplicate files in the concurrent workspaces.

This fix removes the dependency on the node 14 test job for publish to resolve that error.