Remove the Service Metrics alert
This removes the annoying alert when a service is not recognised. This
causes a lot of churn and there's no need for it. next-metrics now
guesses the service name based on the request URL's host name
(https://github.com/Financial-Times/next-metrics/pull/583) and so this
alert will never fire.

This also allows us to do a lot of simplification because we don't have
to handle undefined services.

For all the removed things (e.g. types/options):
a) these was never part of the public API for n-express
b) we don't seem to be using them anywhere

See-also: https://financialtimes.atlassian.net/browse/CPREL-1123