Remove the Heroku log drain check
We're moving this check. We currently have Problem Hub problems based on
the Biz Ops data for Heroku apps, e.g.
https://github.com/Financial-Times/problem-hub/blob/main/problemDefinitions/HerokuApps/herokuAppCustomerProductsMissingLogDrain.yaml

We may also add log drain health checks to cp-healthchecks if we
decide it's an issue. Generally though apps don't change that frequently
and problem hub is probably fine for this.