chore!: stop sending metrics
This doesn't fully remove next-metrics to avoid a large migration,
however it does stop n-express from sending any metrics to Graphite by
default. This is a breaking change because the lack of metrics could
trigger alerts for some apps as data will stop being available for
things like `fetch`.

The second part of this work will arrive when we roll out next-metrics
v13, after this PR is merged
https://github.com/Financial-Times/next-metrics/pull/603