feat: derive isUserSubscribed property from ft-user-subscription header
Replace the logic that derives the isUserSubscribed property from the
ft-user-subscription-status header.
We're considering cases where the new header might be missing or have an
incorrect value. In those, we log a warning a fall back to the previous
ft-user-subscription-status header. This will allow us to detect issues
with the new header before removing the existing one.