fix: remove the withAb option entirely
This option does nothing except output a log, it wasn't really
deprecated it was just not working any more. The only potential breaking
change here is if a TypeScript project is referencing this option but
it's unused across the whole FT tech estate.

I searched Splunk for `WITHAB_OPTION_DEPRECATED` over the last quarter
and the only results are next-subscribe and next-retention. These
don't have type checking but I'm gonna go remove the reference anyway.