fix: remove the `EXPRESS_START` warning
This has always been a hack and it's not very useful, just a spammy log
at the wrong level. I don't think it's a breaking change, as far as I
know we're not using this in any monitoring. If an app wants to log on
startup then it's still possible by providing a callback.

Done as part of the n-express audit:
https://financialtimes.atlassian.net/wiki/spaces/CPP/database/8965226525